Refactors ListSuggestion
